Split the Windows and Posix logic into their own files as part of ext_std_network
