+++ hphp/runtime/ext/std/ext_std_network.cpp
@@ -15,17 +15,12 @@
    +----------------------------------------------------------------------+
 */
 #include "hphp/runtime/ext/std/ext_std_network.h"
+#include "hphp/runtime/ext/std/ext_std_network-internal.h"
 
-#ifdef _MSC_VER
-#include <Windns.h>
-#else
 #include <arpa/inet.h>
-#include <arpa/nameser.h>
 #include <netdb.h>
 #include <netinet/in.h>
-#include <resolv.h>
 #include <sys/socket.h>
-#endif
 
 #include <folly/IPAddress.h>
 #include <folly/ScopeGuard.h>
@@ -40,93 +35,11 @@
 #include "hphp/util/lock.h"
 #include "hphp/util/network.h"
 
-#if defined(__APPLE__)
-# include <arpa/nameser_compat.h>
-#include <vector>
-#endif
-
-// HOST_NAME_MAX is recommended by POSIX, but not required.
-// FreeBSD and OSX (as of 10.9) are known to not define it.
-// 255 is generally the safe value to assume and upstream
-// PHP does this as well.
-#ifndef HOST_NAME_MAX
-#define HOST_NAME_MAX 255
-#endif
-
-#define MAXPACKET  8192 /* max packet size used internally by BIND */
-#define DNS_T_A 1
-#define DNS_T_NS 2
-#define DNS_T_CNAME 5
-#define DNS_T_SOA 6
-#define DNS_T_PTR 12
-#define DNS_T_HINFO 13
-#define DNS_T_MINFO 14
-#define DNS_T_MX 15
-#define DNS_T_TXT 16
-#define DNS_T_AAAA 28
-#define DNS_T_SRV 33
-#define DNS_T_NAPTR 35
-#define DNS_T_A6 38
-#define DNS_T_ANY 255
-
-#define PHP_DNS_NUM_TYPES   12  // Number of DNS Types Supported by PHP
-#define PHP_DNS_A      0x00000001
-#define PHP_DNS_NS     0x00000002
-#define PHP_DNS_CNAME  0x00000010
-#define PHP_DNS_SOA    0x00000020
-#define PHP_DNS_PTR    0x00000800
-#define PHP_DNS_HINFO  0x00001000
-#define PHP_DNS_MX     0x00004000
-#define PHP_DNS_TXT    0x00008000
-#define PHP_DNS_A6     0x01000000
-#define PHP_DNS_SRV    0x02000000
-#define PHP_DNS_NAPTR  0x04000000
-#define PHP_DNS_AAAA   0x08000000
-#define PHP_DNS_ANY    0x10000000
-#define PHP_DNS_ALL    (PHP_DNS_A|PHP_DNS_NS|PHP_DNS_CNAME|PHP_DNS_SOA| \
-                        PHP_DNS_PTR|PHP_DNS_HINFO|PHP_DNS_MX|PHP_DNS_TXT| \
-                        PHP_DNS_A6|PHP_DNS_SRV|PHP_DNS_NAPTR|PHP_DNS_AAAA)
-
 namespace HPHP {
 ///////////////////////////////////////////////////////////////////////////////
 // DNS
 
 static Mutex NetworkMutex;
-
-#ifndef _MSC_VER
-class ResolverInit {
-public:
-  ResolverInit() : m_res(nullptr) {
-    m_res = (struct __res_state *)calloc(1, sizeof(*m_res));
-    initRes();
-  }
-  ~ResolverInit() {
-    if (m_res)
-      free(m_res);
-    m_res = nullptr;
-  }
-
-  struct __res_state *getResolver(void) {
-    initRes();
-    return m_res;
-  }
-
-  static DECLARE_THREAD_LOCAL(ResolverInit, s_res);
-private:
-  void initRes(void) {
-    if (m_res) {
-      memset(m_res, 0, sizeof(*m_res));
-      if (res_ninit(m_res)) {
-        free(m_res);
-        m_res = nullptr;
-      }
-    }
-  }
-
-  struct __res_state *m_res;
-};
-IMPLEMENT_THREAD_LOCAL(ResolverInit, ResolverInit::s_res);
-#endif
 
 Variant HHVM_FUNCTION(gethostname) {
   char h_name[HOST_NAME_MAX];
@@ -301,970 +214,6 @@ String HHVM_FUNCTION(long2ip, const String& proper_address) {
   } catch (folly::IPAddressFormatException &e) {
     return s_empty;
   }
-}
-
-/* just a hack to free resources allocated by glibc in __res_nsend()
- * See also:
- *   res_thread_freeres() in glibc/resolv/res_init.c
- *   __libc_res_nsend()   in resolv/res_send.c
- * */
-
-static void php_dns_free_res(struct __res_state *res) {
-#if defined(__GLIBC__)
-  int ns;
-  for (ns = 0; ns < MAXNS; ns++) {
-    if (res->_u._ext.nsaddrs[ns] != NULL) {
-      free(res->_u._ext.nsaddrs[ns]);
-      res->_u._ext.nsaddrs[ns] = NULL;
-    }
-  }
-#endif
-}
-
-bool HHVM_FUNCTION(checkdnsrr, const String& host,
-                               const String& type /* = null_string */) {
-  IOStatusHelper io("dns_check_record", host.data());
-  const char *stype;
-  if (type.empty()) {
-    stype = "MX";
-  } else {
-    stype = type.data();
-  }
-  if (host.empty()) {
-    throw_invalid_argument("host: [empty]");
-  }
-
-  int ntype;
-  if (!strcasecmp("A", stype)) ntype = DNS_T_A;
-  else if (!strcasecmp("NS",    stype)) ntype = DNS_T_NS;
-  else if (!strcasecmp("MX",    stype)) ntype = DNS_T_MX;
-  else if (!strcasecmp("PTR",   stype)) ntype = DNS_T_PTR;
-  else if (!strcasecmp("ANY",   stype)) ntype = DNS_T_ANY;
-  else if (!strcasecmp("SOA",   stype)) ntype = DNS_T_SOA;
-  else if (!strcasecmp("TXT",   stype)) ntype = DNS_T_TXT;
-  else if (!strcasecmp("CNAME", stype)) ntype = DNS_T_CNAME;
-  else if (!strcasecmp("AAAA",  stype)) ntype = DNS_T_AAAA;
-  else if (!strcasecmp("SRV",   stype)) ntype = DNS_T_SRV;
-  else if (!strcasecmp("NAPTR", stype)) ntype = DNS_T_NAPTR;
-  else if (!strcasecmp("A6",    stype)) ntype = DNS_T_A6;
-  else {
-    throw_invalid_argument("type: %s", stype);
-    return false;
-  }
-
-
-#ifdef _MSC_VER
-  DNS_STATUS      status;                 /* Return value of DnsQuery_A() function */
-  PDNS_RECORD     pResult;          /* Pointer to DNS_RECORD structure */
-  status = DnsQuery_A(host.c_str(), ntype, DNS_QUERY_STANDARD, NULL, &pResult, NULL);
-
-  if (status) {
-    return false;
-  }
-
-  return true;
-#else
-  unsigned char ans[MAXPACKET];
-  struct __res_state *res;
-  res = ResolverInit::s_res.get()->getResolver();
-  if (res == NULL) {
-    return false;
-  }
-
-  int i = res_nsearch(res, host.data(), C_IN, ntype, ans, sizeof(ans));
-  res_nclose(res);
-  php_dns_free_res(res);
-  return (i >= 0);
-#endif
-}
-
-const StaticString
-  s_host("host"),
-  s_type("type"),
-  s_ip("ip"),
-  s_pri("pri"),
-  s_weight("weight"),
-  s_port("port"),
-  s_order("order"),
-  s_pref("pref"),
-  s_target("target"),
-  s_cpu("cpu"),
-  s_os("os"),
-  s_txt("txt"),
-  s_mname("mname"),
-  s_rname("rname"),
-  s_serial("serial"),
-  s_refresh("refresh"),
-  s_retry("retry"),
-  s_expire("expire"),
-  s_minimum_ttl("minimum-ttl"),
-  s_ipv6("ipv6"),
-  s_masklen("masklen"),
-  s_chain("chain"),
-  s_flags("flags"),
-  s_services("services"),
-  s_regex("regex"),
-  s_replacement("replacement"),
-  s_class("class"),
-  s_ttl("ttl"),
-  s_A("A"),
-  s_MX("MX"),
-  s_CNAME("CNAME"),
-  s_NS("NS"),
-  s_PTR("PTR"),
-  s_HINFO("HINFO"),
-  s_TXT("TXT"),
-  s_SOA("SOA"),
-  s_AAAA("AAAA"),
-  s_A6("A6"),
-  s_SRV("SRV"),
-  s_NAPTR("NAPTR"),
-  s_IN("IN");
-
-#ifdef _MSC_VER
-static void php_parserr(PDNS_RECORD pRec, int type_to_fetch, int store, Array* subarray)
-{
-  int type;
-  u_long ttl;
-
-  type = pRec->wType;
-  ttl = pRec->dwTtl;
-
-  if (type_to_fetch != DNS_TYPE_ANY && type != type_to_fetch) {
-    return;
-  }
-
-  if (!store) {
-    return;
-  }
-
-  subarray->set(s_host, pRec->pName);
-  subarray->set(s_class, s_IN);
-  subarray->set(s_ttl, (long)ttl);
-
-  switch (type) {
-  case DNS_TYPE_A: {
-    IN_ADDR ipaddr;
-    ipaddr.S_un.S_addr = (pRec->Data.A.IpAddress);
-    subarray->set(s_type, s_A);
-    subarray->set(s_ip, inet_ntoa(ipaddr));
-    break;
-  }
-
-  case DNS_TYPE_MX:
-    subarray->set(s_type, s_MX);
-    subarray->set(s_pri, pRec->Data.Srv.wPriority);
-    /* no break; */
-
-  case DNS_TYPE_CNAME:
-    if (type == DNS_TYPE_CNAME) {
-      subarray->set(s_type, s_CNAME);
-    }
-    /* no break; */
-
-  case DNS_TYPE_NS:
-    if (type == DNS_TYPE_NS) {
-      subarray->set(s_type, s_NS);
-    }
-    /* no break; */
-
-  case DNS_TYPE_PTR:
-    if (type == DNS_TYPE_PTR) {
-      subarray->set(s_type, s_PTR);
-    }
-    subarray->set(s_target, pRec->Data.MX.pNameExchange);
-    break;
-
-    /* Not available on windows, the query is possible but there is no DNS_HINFO_DATA structure */
-  case DNS_TYPE_HINFO:
-  case DNS_TYPE_TEXT:
-  {
-    DWORD i = 0;
-    DNS_TXT_DATA *data_txt = &pRec->Data.TXT;
-    DWORD count = data_txt->dwStringCount;
-    String txt;
-
-    subarray->set(s_type, s_TXT);
-    for (i = 0; i < count; i++) {
-      txt += String(data_txt->pStringArray[i], CopyString);
-    }
-    subarray->set(s_txt, txt);
-  }
-  break;
-
-  case DNS_TYPE_SOA:
-  {
-    DNS_SOA_DATA *data_soa = &pRec->Data.Soa;
-
-    subarray->set(s_type, s_SOA);
-
-    subarray->set(s_mname, data_soa->pNamePrimaryServer);
-    subarray->set(s_rname, data_soa->pNameAdministrator);
-    subarray->set(s_serial, (int)data_soa->dwSerialNo);
-    subarray->set(s_refresh, (int)data_soa->dwRefresh);
-    subarray->set(s_retry, (int)data_soa->dwRetry);
-    subarray->set(s_expire, (int)data_soa->dwExpire);
-    subarray->set(s_minimum_ttl, (int)data_soa->dwDefaultTtl);
-  }
-  break;
-
-  case DNS_TYPE_AAAA:
-  {
-    DNS_AAAA_DATA *data_aaaa = &pRec->Data.AAAA;
-    char buf[sizeof("AAAA:AAAA:AAAA:AAAA:AAAA:AAAA:AAAA:AAAA")];
-    char *tp = buf;
-    int i;
-    unsigned short out[8];
-    int have_v6_break = 0, in_v6_break = 0;
-
-    for (i = 0; i < 4; ++i) {
-      DWORD chunk = data_aaaa->Ip6Address.IP6Dword[i];
-      out[i * 2] = htons(LOWORD(chunk));
-      out[i * 2 + 1] = htons(HIWORD(chunk));
-    }
-
-    for (i = 0; i < 8; i++) {
-      if (out[i] != 0) {
-        if (tp > buf) {
-          in_v6_break = 0;
-          tp[0] = ':';
-          tp++;
-        }
-        tp += sprintf((char*)tp, "%x", out[i]);
-      }
-      else {
-        if (!have_v6_break) {
-          have_v6_break = 1;
-          in_v6_break = 1;
-          tp[0] = ':';
-          tp++;
-        }
-        else if (!in_v6_break) {
-          tp[0] = ':';
-          tp++;
-          tp[0] = '0';
-          tp++;
-        }
-      }
-    }
-
-    if (have_v6_break && in_v6_break) {
-      tp[0] = ':';
-      tp++;
-    }
-    tp[0] = '\0';
-
-    subarray->set(s_type, s_AAAA);
-    subarray->set(s_ipv6, buf);
-  }
-  break;
-
-  case DNS_TYPE_SRV:
-  {
-    DNS_SRV_DATA *data_srv = &pRec->Data.Srv;
-
-    subarray->set(s_type, s_SRV);
-    subarray->set(s_pri, data_srv->wPriority);
-    subarray->set(s_weight, data_srv->wWeight);
-    subarray->set(s_port, data_srv->wPort);
-    subarray->set(s_target, data_srv->pNameTarget);
-  }
-  break;
-
-  case DNS_TYPE_NAPTR:
-  {
-    DNS_NAPTR_DATA * data_naptr = &pRec->Data.Naptr;
-
-    subarray->set(s_type, s_NAPTR);
-    subarray->set(s_order, data_naptr->wOrder);
-    subarray->set(s_pref, data_naptr->wPreference);
-    subarray->set(s_flags, data_naptr->pFlags);
-    subarray->set(s_services, data_naptr->pService);
-    subarray->set(s_regex, data_naptr->pRegularExpression);
-    subarray->set(s_replacement, data_naptr->pReplacement);
-  }
-  break;
-
-  default:
-    /* unknown type */
-    subarray->clear();
-    return;
-  }
-
-}
-#else
-typedef union {
-  HEADER qb1;
-  u_char qb2[65536];
-} querybuf;
-
-#define CHECKCP(n) do { \
-  if ((cp + (n)) > end) { \
-    return nullptr; \
-  } \
-} while (0)
-
-static unsigned char *php_parserr(unsigned char *cp, unsigned char* end,
-                                  querybuf *answer,
-                                  int type_to_fetch, bool store,
-                                  Array &subarray) {
-  unsigned short type, cls ATTRIBUTE_UNUSED, dlen;
-  unsigned long ttl;
-  int64_t n, i;
-  unsigned short s;
-  unsigned char *tp, *p;
-  char name[255 + 2];  // IETF STD 13 section 3.1; 255 bytes
-  int have_v6_break = 0, in_v6_break = 0;
-
-  n = dn_expand(answer->qb2, answer->qb2+65536, cp, name, sizeof(name) - 2);
-  if (n < 0) {
-    return NULL;
-  }
-  cp += n;
-
-  CHECKCP(10);
-  GETSHORT(type, cp);
-  GETSHORT(cls, cp);
-  GETLONG(ttl, cp);
-  GETSHORT(dlen, cp);
-  CHECKCP(dlen);
-  if (type_to_fetch != T_ANY && type != type_to_fetch) {
-    cp += dlen;
-    return cp;
-  }
-
-  if (!store) {
-    cp += dlen;
-    return cp;
-  }
-
-  subarray.set(s_host, String(name, CopyString));
-  switch (type) {
-  case DNS_T_A:
-    CHECKCP(4);
-    subarray.set(s_type, s_A);
-    snprintf(name, sizeof(name), "%d.%d.%d.%d", cp[0], cp[1], cp[2], cp[3]);
-    subarray.set(s_ip, String(name, CopyString));
-    cp += dlen;
-    break;
-  case DNS_T_MX:
-    CHECKCP(2);
-    subarray.set(s_type, s_MX);
-    GETSHORT(n, cp);
-    subarray.set(s_pri, n);
-    /* no break; */
-  case DNS_T_CNAME:
-    if (type == DNS_T_CNAME) {
-      subarray.set(s_type, s_CNAME);
-    }
-    /* no break; */
-  case DNS_T_NS:
-    if (type == DNS_T_NS) {
-      subarray.set(s_type, s_NS);
-    }
-    /* no break; */
-  case DNS_T_PTR:
-    if (type == DNS_T_PTR) {
-      subarray.set(s_type, s_PTR);
-    }
-    n = dn_expand(answer->qb2, answer->qb2+65536, cp, name, (sizeof name) - 2);
-    if (n < 0) {
-      return NULL;
-    }
-    cp += n;
-    subarray.set(s_target, String(name, CopyString));
-    break;
-  case DNS_T_HINFO:
-    /* See RFC 1010 for values */
-    subarray.set(s_type, s_HINFO);
-    CHECKCP(1);
-    n = *cp & 0xFF;
-    cp++;
-    CHECKCP(n);
-    subarray.set(s_cpu, String((const char *)cp, n, CopyString));
-    cp += n;
-    CHECKCP(1);
-    n = *cp & 0xFF;
-    cp++;
-    CHECKCP(n);
-    subarray.set(s_os, String((const char *)cp, n, CopyString));
-    cp += n;
-    break;
-  case DNS_T_TXT: {
-    int l1 = 0, l2 = 0;
-
-    String s = String(dlen, ReserveString);
-    tp = (unsigned char *)s.mutableData();
-
-    while (l1 < dlen) {
-      n = cp[l1];
-      if ((n + l1) > dlen) {
-        // bad record, don't set anything
-        break;
-      }
-      memcpy(tp + l1 , cp + l1 + 1, n);
-      l1 = l1 + n + 1;
-      l2 = l2 + n;
-    }
-    s.setSize(l2);
-    cp += dlen;
-
-    subarray.set(s_type, s_TXT);
-    subarray.set(s_txt, s);
-    break;
-  }
-  case DNS_T_SOA:
-    subarray.set(s_type, s_SOA);
-    n = dn_expand(answer->qb2, end, cp, name, (sizeof name) -2);
-    if (n < 0) {
-      return NULL;
-    }
-    cp += n;
-    subarray.set(s_mname, String(name, CopyString));
-    n = dn_expand(answer->qb2, end, cp, name, (sizeof name) -2);
-    if (n < 0) {
-      return NULL;
-    }
-    cp += n;
-    subarray.set(s_rname, String(name, CopyString));
-    CHECKCP(5*4);
-    GETLONG(n, cp);
-    subarray.set(s_serial, n);
-    GETLONG(n, cp);
-    subarray.set(s_refresh, n);
-    GETLONG(n, cp);
-    subarray.set(s_retry, n);
-    GETLONG(n, cp);
-    subarray.set(s_expire, n);
-    GETLONG(n, cp);
-    subarray.set(s_minimum_ttl, n);
-    break;
-  case DNS_T_AAAA:
-    tp = (unsigned char *)name;
-    CHECKCP(8*2);
-    for (i = 0; i < 8; i++) {
-      GETSHORT(s, cp);
-      if (s != 0) {
-        if (tp > (u_char *)name) {
-          in_v6_break = 0;
-          tp[0] = ':';
-          tp++;
-        }
-        tp += sprintf((char *)tp, "%x", s);
-      } else {
-        if (!have_v6_break) {
-          have_v6_break = 1;
-          in_v6_break = 1;
-          tp[0] = ':';
-          tp++;
-        } else if (!in_v6_break) {
-          tp[0] = ':';
-          tp++;
-          tp[0] = '0';
-          tp++;
-        }
-      }
-    }
-    if (have_v6_break && in_v6_break) {
-      tp[0] = ':';
-      tp++;
-    }
-    tp[0] = '\0';
-    subarray.set(s_type, s_AAAA);
-    subarray.set(s_ipv6, String(name, CopyString));
-    break;
-  case DNS_T_A6:
-    p = cp;
-    subarray.set(s_type, s_A6);
-    CHECKCP(1);
-    n = ((int)cp[0]) & 0xFF;
-    cp++;
-    subarray.set(s_masklen, n);
-    tp = (unsigned char *)name;
-    if (n > 15) {
-      have_v6_break = 1;
-      in_v6_break = 1;
-      tp[0] = ':';
-      tp++;
-    }
-    if (n % 16 > 8) {
-      /* Partial short */
-      if (cp[0] != 0) {
-        if (tp > (u_char *)name) {
-          in_v6_break = 0;
-          tp[0] = ':';
-          tp++;
-        }
-        sprintf((char *)tp, "%x", cp[0] & 0xFF);
-      } else {
-        if (!have_v6_break) {
-          have_v6_break = 1;
-          in_v6_break = 1;
-          tp[0] = ':';
-          tp++;
-        } else if (!in_v6_break) {
-          tp[0] = ':';
-          tp++;
-          tp[0] = '0';
-          tp++;
-        }
-      }
-      cp++;
-    }
-    for (i = (n + 8)/16; i < 8; i++) {
-      CHECKCP(2);
-      GETSHORT(s, cp);
-      if (s != 0) {
-        if (tp > (u_char *)name) {
-          in_v6_break = 0;
-          tp[0] = ':';
-          tp++;
-        }
-        tp += sprintf((char*)tp,"%x",s);
-      } else {
-        if (!have_v6_break) {
-          have_v6_break = 1;
-          in_v6_break = 1;
-          tp[0] = ':';
-          tp++;
-        } else if (!in_v6_break) {
-          tp[0] = ':';
-          tp++;
-          tp[0] = '0';
-          tp++;
-        }
-      }
-    }
-    if (have_v6_break && in_v6_break) {
-      tp[0] = ':';
-      tp++;
-    }
-    tp[0] = '\0';
-    subarray.set(s_ipv6, String(name, CopyString));
-    if (cp < p + dlen) {
-      n = dn_expand(answer->qb2, end, cp, name,
-                    (sizeof name) - 2);
-      if (n < 0) {
-        return NULL;
-      }
-      cp += n;
-      subarray.set(s_chain, String(name, CopyString));
-    }
-    break;
-  case DNS_T_SRV:
-    CHECKCP(3*2);
-    subarray.set(s_type, s_SRV);
-    GETSHORT(n, cp);
-    subarray.set(s_pri, n);
-    GETSHORT(n, cp);
-    subarray.set(s_weight, n);
-    GETSHORT(n, cp);
-    subarray.set(s_port, n);
-    n = dn_expand(answer->qb2, end, cp, name, (sizeof name) - 2);
-    if (n < 0) {
-      return NULL;
-    }
-    cp += n;
-    subarray.set(s_target, String(name, CopyString));
-    break;
-  case DNS_T_NAPTR:
-    CHECKCP(2*2);
-    subarray.set(s_type, s_NAPTR);
-    GETSHORT(n, cp);
-    subarray.set(s_order, n);
-    GETSHORT(n, cp);
-    subarray.set(s_pref, n);
-
-    CHECKCP(1);
-    n = (cp[0] & 0xFF);
-    ++cp;
-    CHECKCP(n);
-    subarray.set(s_flags, String((const char *)cp, n, CopyString));
-    cp += n;
-
-    CHECKCP(1);
-    n = (cp[0] & 0xFF);
-    ++cp;
-    CHECKCP(n);
-    subarray.set(s_services, String((const char *)cp, n, CopyString));
-    cp += n;
-
-    CHECKCP(1);
-    n = (cp[0] & 0xFF);
-    ++cp;
-    CHECKCP(n);
-    subarray.set(s_regex, String((const char *)cp, n, CopyString));
-    cp += n;
-
-    n = dn_expand(answer->qb2, end, cp, name, (sizeof name) - 2);
-    if (n < 0) {
-      return NULL;
-    }
-    cp += n;
-    subarray.set(s_replacement, String(name, CopyString));
-    break;
-  default:
-    cp += dlen;
-  }
-
-  subarray.set(s_class, s_IN);
-  subarray.set(s_ttl, (int)ttl);
-  return cp;
-}
-#endif
-
-Variant HHVM_FUNCTION(dns_get_record, const String& hostname, int type /*= -1*/,
-                      VRefParam authnsRef /* = null */,
-                      VRefParam addtlRef /* = null */) {
-  IOStatusHelper io("dns_get_record", hostname.data(), type);
-  if (type < 0) type = PHP_DNS_ALL;
-  if (type & ~PHP_DNS_ALL && type != PHP_DNS_ANY) {
-    raise_warning("Type '%d' not supported", type);
-    return false;
-  }
-
-  Array ret;
-
-#ifdef _MSC_VER
-  int type2;
-  Array authns;
-  Array addtl;
-  if (type == PHP_DNS_ANY) {
-    type2 = PHP_DNS_NUM_TYPES + 1;
-  }
-  else {
-    type2 = 0;
-  }
-
-  bool first_query = true;
-  bool store_results = true;
-  for (;
-  type2 < (!addtlRef.isNull() ? (PHP_DNS_NUM_TYPES + 2) : PHP_DNS_NUM_TYPES) || first_query;
-    type2++
-    ) {
-    first_query = false;
-    int type_to_fetch;
-    DNS_STATUS      status;                 /* Return value of DnsQuery_A() function */
-    PDNS_RECORD     pResult, pRec;          /* Pointer to DNS_RECORD structure */
-
-    switch (type2) {
-    case -1: /* raw */
-      type_to_fetch = type;
-      /* skip over the rest and go directly to additional records */
-      type = PHP_DNS_NUM_TYPES - 1;
-      break;
-    case 0:
-      type_to_fetch = type&PHP_DNS_A ? DNS_TYPE_A : 0;
-      break;
-    case 1:
-      type_to_fetch = type&PHP_DNS_NS ? DNS_TYPE_NS : 0;
-      break;
-    case 2:
-      type_to_fetch = type&PHP_DNS_CNAME ? DNS_TYPE_CNAME : 0;
-      break;
-    case 3:
-      type_to_fetch = type&PHP_DNS_SOA ? DNS_TYPE_SOA : 0;
-      break;
-    case 4:
-      type_to_fetch = type&PHP_DNS_PTR ? DNS_TYPE_PTR : 0;
-      break;
-    case 5:
-      type_to_fetch = type&PHP_DNS_HINFO ? DNS_TYPE_HINFO : 0;
-      break;
-    case 6:
-      type_to_fetch = type&PHP_DNS_MX ? DNS_TYPE_MX : 0;
-      break;
-    case 7:
-      type_to_fetch = type&PHP_DNS_TXT ? DNS_TYPE_TEXT : 0;
-      break;
-    case 8:
-      type_to_fetch = type&PHP_DNS_AAAA ? DNS_TYPE_AAAA : 0;
-      break;
-    case 9:
-      type_to_fetch = type&PHP_DNS_SRV ? DNS_TYPE_SRV : 0;
-      break;
-    case 10:
-      type_to_fetch = type&PHP_DNS_NAPTR ? DNS_TYPE_NAPTR : 0;
-      break;
-    case 11:
-      type_to_fetch = type&PHP_DNS_A6 ? DNS_TYPE_A6 : 0;
-      break;
-    case PHP_DNS_NUM_TYPES:
-      store_results = false;
-      continue;
-    default:
-    case (PHP_DNS_NUM_TYPES + 1) :
-      type_to_fetch = DNS_TYPE_ANY;
-      break;
-    }
-
-    if (type_to_fetch) {
-      status = DnsQuery_A(hostname.c_str(), type_to_fetch, DNS_QUERY_STANDARD, NULL, &pResult, NULL);
-
-      if (status) {
-        if (status == DNS_INFO_NO_RECORDS || status == DNS_ERROR_RCODE_NAME_ERROR) {
-          continue;
-        }
-        else {
-          raise_warning("DNS Query failed");
-          return false;
-        }
-      }
-
-      for (pRec = pResult; pRec; pRec = pRec->pNext) {
-        Array retval;
-
-        if (pRec->Flags.S.Section == DnsSectionAnswer) {
-          php_parserr(pRec, type_to_fetch, store_results, &retval);
-          if (!retval.empty() && store_results) {
-            ret.append(retval);
-          }
-        }
-
-        if (!authns.isNull() && pRec->Flags.S.Section == DnsSectionAuthority) {
-
-          php_parserr(pRec, type_to_fetch, 1, &retval);
-          if (!retval.empty()) {
-            authns.append(retval);
-          }
-        }
-
-        /* Stupid typo in PSDK 6.1, WinDNS.h(1258)... */
-#ifndef DnsSectionAdditional
-# ifdef DnsSectionAddtional
-#  define DnsSectionAdditional DnsSectionAddtional
-# else
-# define DnsSectionAdditional 3
-# endif
-#endif
-        if (!addtlRef.isNull() && pRec->Flags.S.Section == DnsSectionAdditional) {
-          php_parserr(pRec, type_to_fetch, 1, &retval);
-          if (!retval.empty()) {
-            addtl.append(retval);
-          }
-        }
-      }
-      /* Free memory allocated for DNS records. */
-      DnsRecordListFree(pResult, DnsFreeRecordListDeep);
-    }
-  }
-  authnsRef = authns;
-  addtlRef = addtl;
-  return ret;
-#else
-  unsigned char *cp = NULL, *end = NULL;
-  int qd, an, ns = 0, ar = 0;
-  querybuf answer;
-
-  /* - We emulate an or'ed type mask by querying type by type.
-   *   (Steps 0 - NUMTYPES-1 )
-   *   If additional info is wanted we check again with DNS_T_ANY
-   *   (step NUMTYPES / NUMTYPES+1 )
-   *   store_results is used to skip storing the results retrieved in step
-   *   NUMTYPES+1 when results were already fetched.
-   * - In case of PHP_DNS_ANY we use the directly fetch DNS_T_ANY.
-   *   (step NUMTYPES+1 )
-   */
-  bool first_query = true;
-  bool store_results = true;
-  for (int t = (type == PHP_DNS_ANY ? (PHP_DNS_NUM_TYPES + 1) : 0);
-       t < PHP_DNS_NUM_TYPES + 2 || first_query; t++) {
-    first_query = false;
-    int type_to_fetch;
-    switch (t) {
-    case 0:  type_to_fetch = type & PHP_DNS_A     ? DNS_T_A     : 0; break;
-    case 1:  type_to_fetch = type & PHP_DNS_NS    ? DNS_T_NS    : 0; break;
-    case 2:  type_to_fetch = type & PHP_DNS_CNAME ? DNS_T_CNAME : 0; break;
-    case 3:  type_to_fetch = type & PHP_DNS_SOA   ? DNS_T_SOA   : 0; break;
-    case 4:  type_to_fetch = type & PHP_DNS_PTR   ? DNS_T_PTR   : 0; break;
-    case 5:  type_to_fetch = type & PHP_DNS_HINFO ? DNS_T_HINFO : 0; break;
-    case 6:  type_to_fetch = type & PHP_DNS_MX    ? DNS_T_MX    : 0; break;
-    case 7:  type_to_fetch = type & PHP_DNS_TXT   ? DNS_T_TXT   : 0; break;
-    case 8:  type_to_fetch = type & PHP_DNS_AAAA  ? DNS_T_AAAA  : 0; break;
-    case 9:  type_to_fetch = type & PHP_DNS_SRV   ? DNS_T_SRV   : 0; break;
-    case 10: type_to_fetch = type & PHP_DNS_NAPTR ? DNS_T_NAPTR : 0; break;
-    case 11: type_to_fetch = type & PHP_DNS_A6    ? DNS_T_A6    : 0; break;
-    case PHP_DNS_NUM_TYPES:
-      store_results = false;
-      continue;
-    default:
-    case (PHP_DNS_NUM_TYPES + 1):
-      type_to_fetch = DNS_T_ANY;
-      break;
-    }
-    if (!type_to_fetch) continue;
-
-    struct __res_state *res;
-    res = ResolverInit::s_res.get()->getResolver();
-    if (res == NULL) {
-      return false;
-    }
-
-    int n = res_nsearch(res, hostname.data(), C_IN, type_to_fetch,
-                        answer.qb2, sizeof answer);
-    if (n < 0) {
-      res_nclose(res);
-      php_dns_free_res(res);
-      continue;
-    }
-
-    HEADER *hp;
-    cp = answer.qb2 + HFIXEDSZ;
-    end = answer.qb2 + n;
-    hp = (HEADER *)&answer;
-    qd = ntohs(hp->qdcount);
-    an = ntohs(hp->ancount);
-    ns = ntohs(hp->nscount);
-    ar = ntohs(hp->arcount);
-
-    /* Skip QD entries, they're only used by dn_expand later on */
-    while (qd-- > 0) {
-      n = dn_skipname(cp, end);
-      if (n < 0) {
-        raise_warning("Unable to parse DNS data received");
-        res_nclose(res);
-        php_dns_free_res(res);
-        return false;
-      }
-      cp += n + QFIXEDSZ;
-    }
-
-    /* YAY! Our real answers! */
-    while (an-- && cp && cp < end) {
-      Array retval;
-      cp = php_parserr(cp, end, &answer, type_to_fetch, store_results, retval);
-      if (!retval.empty() && store_results) {
-        ret.append(retval);
-      }
-    }
-    res_nclose(res);
-    php_dns_free_res(res);
-  }
-
-  Array authns;
-  Array addtl;
-
-  /* List of Authoritative Name Servers */
-  while (ns-- > 0 && cp && cp < end) {
-    Array retval;
-    cp = php_parserr(cp, end, &answer, DNS_T_ANY, true, retval);
-    if (!retval.empty()) {
-      authns.append(retval);
-    }
-  }
-
-  /* Additional records associated with authoritative name servers */
-  while (ar-- > 0 && cp && cp < end) {
-    Array retval;
-    cp = php_parserr(cp, end, &answer, DNS_T_ANY, true, retval);
-    if (!retval.empty()) {
-      addtl.append(retval);
-    }
-  }
-
-  authnsRef.assignIfRef(authns);
-  addtlRef.assignIfRef(addtl);
-  return ret;
-#endif
-}
-
-bool HHVM_FUNCTION(getmxrr, const String& hostname,
-                            VRefParam mxhostsRef,
-                            VRefParam weightsRef /* = null */) {
-  IOStatusHelper io("dns_get_mx", hostname.data());
-  Array mxhosts;
-  Array weights;
-  SCOPE_EXIT {
-    mxhostsRef.assignIfRef(mxhosts);
-    weightsRef.assignIfRef(weights);
-  };
-#ifdef _MSC_VER
-  DNS_STATUS      status;                 /* Return value of DnsQuery_A() function */
-  PDNS_RECORD     pResult, pRec;          /* Pointer to DNS_RECORD structure */
-
-  status = DnsQuery_A(hostname.c_str(), DNS_TYPE_MX, DNS_QUERY_STANDARD, NULL, &pResult, NULL);
-
-  if (status) {
-    return false;
-  }
-
-  for (pRec = pResult; pRec; pRec = pRec->pNext) {
-    DNS_SRV_DATA *srv = &pRec->Data.Srv;
-
-    if (pRec->wType != DNS_TYPE_MX) {
-      continue;
-    }
-
-    mxhosts.append(String(pRec->Data.MX.pNameExchange, CopyString));
-    weights.append(srv->wPriority);
-  }
-
-  /* Free memory allocated for DNS records. */
-  DnsRecordListFree(pResult, DnsFreeRecordListDeep);
-  return true;
-#else
-  int count, qdc;
-  unsigned short type, weight;
-  unsigned char ans[MAXPACKET];
-  char buf[255 + 1];  // IETF STD 13 section 3.1; 255 bytes
-  unsigned char *cp, *end;
-
-  /* Go! */
-  struct __res_state *res;
-  res = ResolverInit::s_res.get()->getResolver();
-  if (res == NULL) {
-    return false;
-  }
-
-  int i = res_nsearch(res, hostname.data(), C_IN, DNS_T_MX,
-                      (unsigned char*)&ans, sizeof(ans));
-  if (i < 0) {
-    res_nclose(res);
-    php_dns_free_res(res);
-    return false;
-  }
-  if (i > (int)sizeof(ans)) {
-    i = sizeof(ans);
-  }
-  HEADER *hp = (HEADER *)&ans;
-  cp = (unsigned char *)&ans + HFIXEDSZ;
-  end = (unsigned char *)&ans +i;
-  for (qdc = ntohs((unsigned short)hp->qdcount); qdc--; cp += i + QFIXEDSZ) {
-    if ((i = dn_skipname(cp, end)) < 0 ) {
-      res_nclose(res);
-      php_dns_free_res(res);
-      return false;
-    }
-  }
-  count = ntohs((unsigned short)hp->ancount);
-  while (--count >= 0 && cp < end) {
-    if ((i = dn_skipname(cp, end)) < 0 ) {
-      res_nclose(res);
-      php_dns_free_res(res);
-      return false;
-    }
-    cp += i;
-    GETSHORT(type, cp);
-    cp += INT16SZ + INT32SZ;
-    GETSHORT(i, cp);
-    if (type != DNS_T_MX) {
-      cp += i;
-      continue;
-    }
-    GETSHORT(weight, cp);
-    if ((i = dn_expand(ans, end, cp, buf, sizeof(buf)-1)) < 0) {
-      res_nclose(res);
-      php_dns_free_res(res);
-      return false;
-    }
-    cp += i;
-    mxhosts.append(String(buf, CopyString));
-    weights.append(weight);
-  }
-  res_nclose(res);
-  php_dns_free_res(res);
-  return true;
-#endif
 }
 
 ///////////////////////////////////////////////////////////////////////////////
